Project Files/requirements.txt
Flask>=3.1.0,<4
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
google-genai>=1.0.0,<2
google-generativeai>=0.8.0,<1
